This is the fix for #12901. I added the check to enusre the search service is on and i moved the method to the AccountRequestSearchIT.java
