+++ src/it/java/teammates/it/storage/sqlsearch/AccountRequestSearchIT.java
@@ -15,6 +15,7 @@ import teammates.storage.sqlentity.AccountRequest;
 import teammates.test.AssertHelper;
 import teammates.test.TestProperties;
 
+
 /**
  * SUT: {@link AccountRequestsDb},
  *      {@link teammates.storage.search.AccountRequestSearchDocument}.
@@ -164,4 +165,26 @@ public class AccountRequestSearchIT extends BaseTestCaseWithSqlDatabaseAccess {
         AssertHelper.assertSameContentIgnoreOrder(Arrays.asList(expected), actual);
     }
 
+    @Test
+    public void testSqlInjectionSearchAccountRequestsInWholeSystem() throws Exception {
+        ______TS("SQL Injection test in searchAccountRequestsInWholeSystem");
+
+        if (!TestProperties.isSearchServiceActive()) { 
+            return; 
+        } 
+
+        AccountRequest accountRequest = new AccountRequest("test@gmail.com", "name", "institute");
+        accountRequestsDb.createAccountRequest(accountRequest);
+
+        String searchInjection = "institute'; DROP TABLE account_requests; --";
+        List<AccountRequest> actualInjection = accountRequestsDb.searchAccountRequestsInWholeSystem(searchInjection);
+        assertEquals(0, actualInjection.size());
+
+        AccountRequest actual = accountRequestsDb.getAccountRequest("test@gmail.com", "institute");
+        assertEquals(accountRequest, actual);
+    }
 }
+
+
+
+
